Fix: EmitRate can be high

+++ src/main/kotlin/model/Geyser.kt
@@ -32,8 +32,13 @@ data class Geyser(
     val x: Short,
     val y: Short,
 
-    /** Emit rate in gram per second when active. */
-    val emitRate: Short,
+    /**
+     * Emit rate in gram per second when active.
+     *
+     * Can be quite high for some geysers.
+     * Like 221888
+     */
+    val emitRate: Int,
 
     /** Average emit rate in gram per second. */
     val avgEmitRate: Short,
